Relocate metrics tracker into plugin reporting
Move createMetricsTracker from src/shared/reporting to src/plugin/src/reporting because only the plugin consumes it. This keeps the helper with the identifier-case and project-index modules that use it instead of routing through shared imports.
